Actual Fixes To SO question implemented
The issue was not in the spacing or bracket balancing. That was okay as I discovered. Though, it was good thought and attempt. The issue appears to relate to pkgIndex.tcl order. For some unidentified reason, placing it as the last file causes the 'xxx does not refer to an object', but moving it up to the middle the 'object/mixin is created. </shrug emoji>