Extra comment details about json_spec with W3C Perform Actions in http_templates
W3C_Perform_Actions in the http_templates_of_webdriver_protocol.tcl can be made to work with a nested TCL dict.
However currently it is using raw json rather than compiling.
In case of future changes (in particular when/if an Actions scheduler) is architected the spec is written for convenience.