Merge out duplicate Exceptions from OLD and W3C
Since Exceptions defined changed in W3C. We can add the new ones that arent covered and remove the old ones with the same name. While now the array has mostly uncovered that are no longer needed in W3C. It doesnt hurt, so we will leave it in. Though if wish to define only W3C Exceptions we can replace the entire array contents (except WebDriverException probably) with the commented out array that was in the last commit